+++ src/app/tests/router-test.js
@@ -197,7 +197,7 @@ routerSuite.add(new Y.Test.Case({
         Y.config.throwFail = this.throwFail;
         delete this.throwFail;
     },
-    
+
     'route() should add a route': function () {
         var router = this.router = new Y.Router();
 
@@ -251,17 +251,17 @@ routerSuite.add(new Y.Test.Case({
         Assert.isTrue(router.hasRoute('/baz?a=b')); //this matches /:foo
         Assert.isFalse(router.hasRoute('/baz/quux'));
         Assert.isFalse(router.hasRoute('/baz/quux?a=b'));
-        
+
         //need to test a router that doesn't have a /:foo catch-all
         router2.route('/foo', noop);
         router2.route('/bar', noop);
-        
+
         Assert.isTrue( router2.hasRoute('/foo'));
         Assert.isTrue( router2.hasRoute('/bar'));
         Assert.isTrue( router2.hasRoute('/bar?a=b'));
         Assert.isFalse(router2.hasRoute('/baz?a=b'));
     },
-    
+
     'hasRoute() should support full URLs': function () {
         var router = this.router = new Y.Router(),
             loc    = win && win.location,
@@ -280,7 +280,7 @@ routerSuite.add(new Y.Test.Case({
         // Scheme-relative URL.
         Assert.isTrue(router.hasRoute('//' + loc.host + '/foo'));
     },
-    
+
     'hasRoute() should always return `false` for URLs with different origins': function () {
         var router = this.router = new Y.Router(),
             origin = 'http://something.really.random.com';
@@ -532,55 +532,55 @@ routerSuite.add(new Y.Test.Case({
 
         router._dispatch('/foo', {}, src);
     },
-    
+
     '_getRegex() should return regexes that do not match too much' : function() {
         var router = this.router = new Y.Router(),
             check = function(path, url) {
                 return router._getRegex(path, []).test(url);
             };
-        
+
         Assert.isTrue( check("/*", "/"));
         Assert.isTrue( check("/*", "/foo"));
         Assert.isTrue( check("/*", "/foo?a=b"));
         Assert.isTrue( check("/*", "/foo#a"));
         Assert.isTrue( check("/*", "/foo/bar"));
-        
+
         Assert.isTrue( check("/*foo", "/"));
         Assert.isTrue( check("/*foo", "/foo"));
         Assert.isTrue( check("/*foo", "/foo?a=b"));
         Assert.isTrue( check("/*foo", "/foo#a"));
         Assert.isTrue( check("/*foo", "/foo/bar"));
-        
+
         Assert.isTrue( check("/", "/"));
         Assert.isFalse(check("/", "/foo"));
         Assert.isFalse(check("/", "/foo/bar"));
         Assert.isFalse(check("/", "/foo?bar"));
         Assert.isFalse(check("/", "/foo#bar"));
-        
+
         Assert.isTrue( check("/foo", "/foo"));
         Assert.isFalse(check("/foo", "/"));
         Assert.isFalse(check("/foo", "/foo/bar"));
         Assert.isFalse(check("/foo", "/foo?bar"));
         Assert.isFalse(check("/foo", "/foo#bar"));
-        
+
         Assert.isTrue( check("/foo/bar", "/foo/bar"));
         Assert.isFalse(check("/foo/bar", "/"));
         Assert.isFalse(check("/foo/bar", "/foo"));
         Assert.isFalse(check("/foo/bar", "/foo?bar"));
         Assert.isFalse(check("/foo/bar", "/foo#bar"));
-        
+
         Assert.isTrue( check("/:foo", "/foo"));
         Assert.isTrue( check("/:foo", "/bar"));
         Assert.isFalse(check("/:foo", "/baz/quux"));
         Assert.isFalse(check("/:foo", "/bar?a=b"));
         Assert.isFalse(check("/:foo", "/bar#a"));
-        
+
         Assert.isTrue( check("/foo/:foo", "/foo/bar"));
         Assert.isTrue( check("/foo/:foo", "/foo/bar"));
         Assert.isFalse(check("/foo/:foo", "/baz/quux"));
         Assert.isFalse(check("/foo/:foo", "/foo/bar?a=b"));
         Assert.isFalse(check("/foo/:foo", "/foo/bar#a"));
-        
+
         Assert.isTrue( check("/:foo/bar", "/foo/bar"));
         Assert.isTrue( check("/:foo/bar", "/bar/bar"));
         Assert.isFalse(check("/:foo/bar", "/baz"));
@@ -774,7 +774,7 @@ routerSuite.add(new Y.Test.Case({
 
         Assert.areSame(2, calls);
     },
-    
+
 
     'routes containing a "*" should match the segments which follow it': function () {
         var calls  = 0,
@@ -801,10 +801,10 @@ routerSuite.add(new Y.Test.Case({
     },
 
     'multiple routers should be able to coexist and have duplicate route handlers': function () {
-        var calls = 0,
+        var calls     = 0,
             routerOne = this.router  = new Y.Router(),
             routerTwo = this.router2 = new Y.Router();
-            
+
         routerOne.route('/baz', function () {
             calls += 1;
         });
@@ -812,11 +812,34 @@ routerSuite.add(new Y.Test.Case({
         routerTwo.route('/baz', function () {
             calls += 1;
         });
-        
+
         routerOne.save('/baz');
 
         this.wait(function () {
             Assert.areSame(2, calls);
+        }, 200);
+    },
+
+    'multiple routers should respond to history events after one router is destroyed': function () {
+        var calls     = 0,
+            routerOne = this.router  = new Y.Router(),
+            routerTwo = this.router2 = new Y.Router();
+
+        routerOne.route('/baz', function () {
+            calls += 1;
+        });
+
+        routerTwo.route('/baz', function () {
+            calls += 1;
+        });
+
+        // Make sure calling `destroy()` doesn't detach `routerTwo`'s history
+        // event listener.
+        routerOne.destroy();
+        routerTwo.save('/baz');
+
+        this.wait(function () {
+            Assert.areSame(1, calls);
         }, 200);
     }
 }));