+++ src/charts/js/AreaSplineSeries.js
@@ -8,7 +8,7 @@
  * @uses Fills
  * @uses CurveUtil
  */
-Y.AreaSplineSeries = Y.Base.create("areaSplineSeries", Y.CartesianSeries, [Y.Fills, Y.CurveUtil], {
+Y.AreaSplineSeries = Y.Base.create("areaSplineSeries", Y.AreaSeries, [Y.CurveUtil], {
     /**
      * @protected
      *