+++ src/yui/tests/assets/good.js
@@ -1,0 +1,3 @@
+YUI.add('good', function(Y) {
+    Y.GOOD = true;
+});